refactor: socket logic from Chat.jsx
